backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
yfinance>=0.2.0
numpy>=1.21
pandas>=1.3
pytest
requests>=2.28
python-dotenv>=0.21
gunicorn>=21.2
scikit-learn>=1.2
ta>=0.10.2